requirements.txt
qiskit>=1.0.0
qiskit-ibm-runtime>=0.20.0
python-dotenv>=1.0.0

